Add Socrata (data.oaklandca.gov) download path as alternative to NetFile
bin/download-socrata fetches all 16 filing sheets from the City's open
data portal and writes CSVs byte-compatible with the NetFile bulk-export
pipeline (same headers, YYYY-MM-DD dates, Socrata-only geocoding columns
dropped). `make download-COAK` uses it; `make download-netfile` keeps
the legacy per-year path.

NOT wired into production: the City's Socrata sync currently drops most
amendments filed since early 2025 (see warning in bin/download-socrata),
so the portal serves stale data for any amended report. Schema fixes
that surfaced during testing (496 Amount was rounded to whole dollars by
the old Excel conversion; B1-Loans Filer_ID too narrow for COAK- local
IDs) are applied here and on master.

+++ dbschema/496.sql
@@ -10,7 +10,7 @@ CREATE TABLE "496" (
 	"Rec_Type" VARCHAR(4) NOT NULL, 
 	"Form_Type" VARCHAR(4) NOT NULL, 
 	"Tran_ID" VARCHAR(12) NOT NULL, 
-	"Amount" INTEGER NOT NULL, 
+	"Amount" FLOAT NOT NULL,
 	"Exp_Date" DATE NOT NULL, 
 	"Date_Thru" VARCHAR(32), 
 	"Expn_Dscr" VARCHAR(90),
